Added a JausMessageEvent to track messages in and out of the Node Manager. These currently come back to the main.cpp file and the EventHandler class but are commented our b/c it creates a lot of spam. Can be used for debug purposes.

+++ openjaus/branches/OpenJAUSv3.2/libopenJAUS/src/nodeManagerInterface/lmHandler.c
@@ -207,7 +207,7 @@ void lmHandlerReceiveLargeMessage(NodeManagerInterface nmi, JausMessage message)
 					{
 						tempMessage = (JausMessage)msgList->messages->elementData[index];
 						index++;
-						if(index > msgList->messages->elementCount)
+						if(index > (unsigned int) msgList->messages->elementCount)
 						{
 							// Invalid set of messages
 							//TODO: Here is when you would request a retransmittal if ACK/NAK is set and ask the sending component to resend some packets
